Basic client-side rtc implementation
