+++ lib/membrane_live_web/templates/page/index.html.heex
@@ -1,1 +1,4 @@
-<div id="main"/>
+<div>
+    <div id="main"/>
+    <video width={1000} height={700} id="123123123123"></video>
+</div>
